Web, Music Player(without song list), Drawer work fine

+++ app/src/main/java/com/damrad/multimediacenter/ui/drawer/MyBrush.kt
@@ -2,4 +2,4 @@ package com.damrad.multimediacenter.ui.drawer
 
 import android.graphics.Path
 
-class MyBrush(var color: Int, var emboss: Boolean, var blur: Boolean, var strokeWidth: Int, var path: Path)
+class MyBrush(var color: Int, var emboss: Boolean, var blur: Boolean, var strokeWidth: Float, var path: Path)
